Rename ReactFizzContext to ReactFizzLegacyContext
